scripts/requirements.txt
PyMuPDF==1.23.8
Pillow==10.2.0
tqdm==4.66.1
